Rename MapboxDirectionsCompat to CMapboxDirections
