+++ Package.swift
@@ -12,8 +12,8 @@ let package = Package(
             targets: ["MapboxDirections"]
         ),
         .library(
-            name: "MapboxDirectionsCompat",
-            targets: ["MapboxDirectionsCompat"]
+            name: "CMapboxDirections",
+            targets: ["CMapboxDirections"]
         ),
     ],
     dependencies: [
@@ -25,9 +25,9 @@ let package = Package(
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
             name: "MapboxDirections",
-            dependencies: ["MapboxDirectionsCompat", "Polyline"]),
+            dependencies: ["CMapboxDirections", "Polyline"]),
         .target(
-            name: "MapboxDirectionsCompat"
+            name: "CMapboxDirections"
         ),
         .testTarget(
             name: "MapboxDirectionsTests",
