+++ Sources/CMapboxDirections/include/MBAttribute.h
