+++ Sources/MapboxDirections/MBRouteOptions.swift
@@ -1,7 +1,7 @@
 import Foundation
 import CoreLocation
 #if SWIFT_PACKAGE
-import MapboxDirectionsCompat
+import CMapboxDirections
 #endif
