+++ Sources/MapboxDirections/MBRouteStep.swift
@@ -2,7 +2,7 @@ import Foundation
 import CoreLocation
 import Polyline
 #if SWIFT_PACKAGE
-import MapboxDirectionsCompat
+import CMapboxDirections
 #endif
