+++ Sources/MapboxDirections/Match/MBMatchOptions.swift
@@ -1,7 +1,7 @@
 import Foundation
 import CoreLocation
 #if SWIFT_PACKAGE
-import MapboxDirectionsCompat
+import CMapboxDirections
 #endif
